+++ test/mpi-cpu.py
@@ -1,12 +1,12 @@
 try:
     import hoomd
-    dev = hoomd.device.CPU
+    dev = hoomd.device.CPU()
     assert(dev.communicator.num_ranks == 2)
 
     if (dev.communicator.rank == 0):
         results = open('test-results-cpu.out', 'a')
         results.write('** Starting MPI CPU tests **\n')
-        results.write('HOOMD version     : {}\n'.format(hoomd.__version__))
+        results.write('HOOMD version     : {}\n'.format(hoomd.version.version))
         results.write('** MPI CPU tests PASSED **\n')
 except:
     with open('test-results-cpu.out', 'a') as results: